backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.4
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
requests==2.31.0
